ex05/requirements.txt
# =====================================================================
# ex05 RAG Q&A 엔진 — Python 의존성 패키지 목록
# =====================================================================
# 설치 방법:
#   python -m venv .venv
#   source .venv/bin/activate   # macOS/Linux
#   .venv\Scripts\activate      # Windows
#   pip install -r requirements.txt
# =====================================================================
# Python 3.10~3.12 권장

# --- LangChain 생태계 ---
langchain==0.3.21
langchain-community==0.3.20
langchain-ollama==0.2.3
langchain-openai==0.3.7
langchain-chroma==0.2.6

# --- 벡터 DB ---
chromadb==1.5.1

# --- 문서 파싱 (ChromaDB 자동 구축용) ---
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5

# --- 임베딩 모델 ---
sentence-transformers==3.3.1

# --- 웹 프레임워크 ---
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20

# --- 환경 변수 ---
python-dotenv==1.0.1

# --- 유틸리티 ---
httpx==0.28.1
pydantic==2.10.6
